refactor: 重构windows截图服务
